backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart